Separated all functions that handle user information from all classes and situation them in User_Info.py
