+++ src/main.py
@@ -1,5 +1,5 @@
-import MainMenu
 import customtkinter 
+import MainMenu
 
 mainmenu = MainMenu.Home()
 mainmenu.mainloop()
